requirements.txt
streamlit
python-binance